% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qobject.R
\name{QObject}
\alias{QObject}
\title{QObject is used to represent C++ Qt objects}
\description{
QObject is used to represent C++ Qt objects

QObject is used to represent C++ Qt objects
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{objectSignals}}{List of callbacks that get invoked upon signal emission}

\item{\code{propertyCache}}{Cache of all properties, updated when a notify signal is emitted}

\item{\code{id}}{The name of object.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{objectSignals}}{List of callbacks that get invoked upon signal emission}

\item{\code{propertyCache}}{Cache of all properties, updated when a notify signal is emitted}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{QObject$new()}}
\item \href{#method-unwrapQObject}{\code{QObject$unwrapQObject()}}
\item \href{#method-unwrapProperties}{\code{QObject$unwrapProperties()}}
\item \href{#method-propertyUpdate}{\code{QObject$propertyUpdate()}}
\item \href{#method-signalEmitted}{\code{QObject$signalEmitted()}}
\item \href{#method-clone}{\code{QObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QObject$new(name, data, webChannel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unwrapQObject"></a>}}
\subsection{Method \code{unwrapQObject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QObject$unwrapQObject(response = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unwrapProperties"></a>}}
\subsection{Method \code{unwrapProperties()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QObject$unwrapProperties()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-propertyUpdate"></a>}}
\subsection{Method \code{propertyUpdate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QObject$propertyUpdate(signals, propertyMap)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-signalEmitted"></a>}}
\subsection{Method \code{signalEmitted()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QObject$signalEmitted(signalName, signalArgs)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
