% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwebchannel.R
\name{QWebChannel}
\alias{QWebChannel}
\title{QWebChannel is used to communicate with QWebChannel C++ Client}
\description{
QWebChannel is used to communicate with QWebChannel C++ Client

QWebChannel is used to communicate with QWebChannel C++ Client
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{transport}}{A R6 object used to transport message.}

\item{\code{execCallbacks}}{List of callbacks which should be executed after receiving data}

\item{\code{execId}}{Identity number of QObject method excuting.}

\item{\code{objects}}{List of QObject which caching object data of C++ side.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{QWebChannel$new()}}
\item \href{#method-exec}{\code{QWebChannel$exec()}}
\item \href{#method-send}{\code{QWebChannel$send()}}
\item \href{#method-debug}{\code{QWebChannel$debug()}}
\item \href{#method-clone}{\code{QWebChannel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a new QWebChannel object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QWebChannel$new(transport, initCallback)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{transport}}{A R6 object with \code{send} method}

\item{\code{initCallback}}{Function to excute after building connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exec"></a>}}
\subsection{Method \code{exec()}}{
Record and send message
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QWebChannel$exec(data, callback)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Message data to send}

\item{\code{callback}}{Function to excute after receiving returned information}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send"></a>}}
\subsection{Method \code{send()}}{
Send data to C++ client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QWebChannel$send(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A string or object which can be serialized to JSON format
Send debug message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-debug"></a>}}
\subsection{Method \code{debug()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QWebChannel$debug(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{Data to send.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QWebChannel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
